+++ mediation/InMobi/CHANGELOG.md
@@ -1,5 +1,9 @@
 ## InMobi Unity Mediation Plugin Changelog
 
+#### Next Version
+- Moved adapter content to `GoogleMobileAds/Mediation/InMobi/`.
+- Refactored adapter namespace to use `GoogleMobileAds.Mediation.InMobi`.
+
 #### [Version 4.2.2](https://dl.google.com/googleadmobadssdk/mediation/unity/inmobi/InMobiUnityAdapter-4.2.2.zip)
 - Supports [InMobi Android adapter version 10.1.2.1](https://github.com/googleads/googleads-mobile-android-mediation/blob/main/ThirdPartyAdapters/inmobi/CHANGELOG.md#version-10121).
 - Supports [InMobi iOS adapter version 10.1.2.1](https://github.com/googleads/googleads-mobile-ios-mediation/blob/main/adapters/InMobi/CHANGELOG.md#version-10121).